Split widget: timeline panel
